icv-project-blip/requirements.txt
# Computer Vision
opencv-python>=4.8.0
ultralytics>=8.0.0
numpy>=1.24.0
Pillow>=10.0.0

# Audio Processing
openai-whisper>=20231117
sounddevice>=0.4.6
librosa>=0.10.0

# Text-to-Speech
pyttsx3>=2.90
pywin32>=305  # Required for Windows COM support (SAPI)

# AI/ML
google-generativeai>=0.3.0  # Used in list_models.py
transformers>=4.30.0
torch>=2.0.0
pillow>=10.0.0
